feat(journal): Enable inline image click to open attachment viewer
Adds functionality to the journal editor to allow users to click on inline generated images. Clicking an image will open the attachment viewer with all inline images from the entry, allowing users to view them in a gallery format. This is triggered by a new click event listener attached to the editor.